requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Automat==25.4.16
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
constantly==23.10.4
cryptography==45.0.5
cssselect==1.2.0
deprecation==2.1.0
filelock==3.18.0
gotrue==2.12.3
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.27.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.11.0
itemloaders==1.3.2
jmespath==1.0.1
lxml==5.1.0
packaging==25.0
parsel==1.8.1
playwright==1.53.0
postgrest==1.1.1
Protego==0.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyDispatcher==2.0.7
pyee==13.0.0
PyJWT==2.10.1
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
queuelib==1.8.0
realtime==2.5.3
requests==2.32.4
requests-file==2.1.0
Scrapy==2.11.1
scrapy-playwright @ git+https://github.com/scrapy-plugins/scrapy-playwright.git@97717e94b9143eaac422295008a0e9e11253a01c
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
storage3==0.12.0
StrEnum==0.4.15
supabase==2.16.0
supafunc==0.10.1
tldextract==5.3.0
Twisted==25.5.0
typing_extensions==4.14.1
urllib3==2.5.0
w3lib==2.3.1
websockets==15.0.1
zope.interface==7.2
